refactor: internal enhancements and reorganization
