+++ bricks/hooks/hooks/post_gen.dart
@@ -1,1 +1,1 @@
-import 'dart:io';void main(){final file=File('.post_gen.txt');file.writeAsStringSync('post_gen: {{name}}');}
+import 'dart:io';import 'package:mason/mason.dart';void run(HookContext context){final file=File('.post_gen.txt');file.writeAsStringSync('post_gen: {{name}}');}
