+++ bricks/hooks/hooks/pre_gen.dart
@@ -1,1 +1,1 @@
-import 'dart:io';void main(){final file=File('.pre_gen.txt');file.writeAsStringSync('pre_gen: {{name}}');}
+import 'dart:io';import 'package:mason/mason.dart';void run(HookContext context){final file=File('.pre_gen.txt');file.writeAsStringSync('pre_gen: {{name}}');}
