+++ packages/mason/lib/src/generator.dart
@@ -12,7 +12,9 @@ import 'package:meta/meta.dart';
 import 'package:path/path.dart' as p;
 import 'package:universal_io/io.dart' show Directory, File, FileMode, Process;
 
-part 'hook_context.dart';
+part 'generator_target.dart';
+part 'hooks.dart';
+part 'mason_generator.dart';
 
 final _partialRegExp = RegExp(r'\{\{~\s(.+)\s\}\}');
 final _fileRegExp = RegExp(r'{{%\s?([a-zA-Z]+)\s?%}}');
@@ -34,215 +36,6 @@ RegExp _loopValueRegExp([String name = '.*?']) {
 
 RegExp _loopInnerRegExp([String name = '.*?']) {
   return RegExp('{{#$name}}(.*?{{{.*?}}}.*?){{/$name}}');
-}
-
-/// {@template mason_generator}
-/// A [MasonGenerator] which extends [Generator] and
-/// exposes the ability to create a [Generator] from a
-/// [Brick].
-/// {@endtemplate}
-class MasonGenerator extends Generator {
-  /// {@macro mason_generator}
-  MasonGenerator(
-    String id,
-    String description, {
-    List<TemplateFile?> files = const <TemplateFile>[],
-    GeneratorHooks hooks = const GeneratorHooks(),
-    this.vars = const <String>[],
-  }) : super(id, description, hooks) {
-    for (final file in files) {
-      addTemplateFile(file);
-    }
-  }
-
-  /// Factory which creates a [MasonGenerator] based on
-  /// a configuration file for a [BrickYaml]:
-  ///
-  /// ```yaml
-  /// name: greetings
-  /// description: A Simple Greetings Template
-  /// vars:
-  ///   - name
-  /// ```
-  static Future<MasonGenerator> fromBrickYaml(BrickYaml brick) async {
-    final brickRoot = File(brick.path!).parent.path;
-    final brickDirectory = p.join(brickRoot, BrickYaml.dir);
-    final brickFiles = Directory(brickDirectory)
-        .listSync(recursive: true)
-        .whereType<File>()
-        .map((file) {
-      return () async {
-        try {
-          final content = await File(file.path).readAsBytes();
-          final relativePath = file.path.substring(
-            file.path.indexOf(BrickYaml.dir) + 1 + BrickYaml.dir.length,
-          );
-          return TemplateFile.fromBytes(relativePath, content);
-        } on Exception {
-          return null;
-        }
-      }();
-    });
-
-    return MasonGenerator(
-      brick.name,
-      brick.description,
-      vars: brick.vars.keys.toList(),
-      files: await Future.wait(brickFiles),
-      hooks: await GeneratorHooks.fromBrickYaml(brick),
-    );
-  }
-
-  /// Factory which creates a [MasonGenerator] based on
-  /// a local [MasonBundle].
-  static Future<MasonGenerator> fromBundle(MasonBundle bundle) async {
-    return MasonGenerator(
-      bundle.name,
-      bundle.description,
-      vars: bundle.vars.keys.toList(),
-      files: _decodeConcatenatedData(bundle.files),
-      hooks: GeneratorHooks.fromBundle(bundle),
-    );
-  }
-
-  /// Factory which creates a [MasonGenerator] based on
-  /// a [GitPath] for a remote [BrickYaml] file.
-  static Future<MasonGenerator> fromGitPath(GitPath gitPath) async {
-    final directory = await BricksJson.temp().add(Brick(git: gitPath));
-    final file = File(p.join(directory, gitPath.path, BrickYaml.file));
-    final brickYaml = checkedYamlDecode(
-      file.readAsStringSync(),
-      (m) => BrickYaml.fromJson(m!),
-    ).copyWith(path: file.path);
-    return MasonGenerator.fromBrickYaml(brickYaml);
-  }
-
-  /// Optional list of variables which will be used to populate
-  /// the corresponding mustache variables within the template.
-  final List<String> vars;
-}
-
-/// Supported types of [GeneratorHooks].
-enum GeneratorHook {
-  /// Script run immediately before the `generate` method is invoked.
-  preGen,
-
-  /// Script run immediately after the `generate` method is invoked.
-  postGen,
-}
-
-extension on GeneratorHook {
-  String toFileName() {
-    switch (this) {
-      case GeneratorHook.preGen:
-        return 'pre_gen.dart';
-      case GeneratorHook.postGen:
-        return 'post_gen.dart';
-    }
-  }
-}
-
-/// {@template generator_hooks}
-/// Scripts that run automatically whenever a particular event occurs
-/// in a [Generator].
-/// {@endtemplate}
-class GeneratorHooks {
-  /// {@macro generator_hooks}
-  const GeneratorHooks({this.preGen, this.postGen, this.pubspec});
-
-  /// Creates [GeneratorHooks] from a provided [MasonBundle].
-  factory GeneratorHooks.fromBundle(MasonBundle bundle) {
-    ScriptFile? _decodeHookScript(MasonBundledFile? file) {
-      if (file == null) return null;
-      final path = file.path;
-      final raw = file.data.replaceAll(_whiteSpace, '');
-      final decoded = base64.decode(raw);
-      try {
-        return ScriptFile.fromBytes(path, decoded);
-      } catch (_) {
-        return null;
-      }
-    }
-
-    List<int>? _decodeHookPubspec(MasonBundledFile? file) {
-      if (file == null) return null;
-      final raw = file.data.replaceAll(_whiteSpace, '');
-      return base64.decode(raw);
-    }
-
-    final preGen = bundle.hooks.firstWhereOrNull(
-      (element) {
-        return p.basename(element.path) == GeneratorHook.preGen.toFileName();
-      },
-    );
-    final postGen = bundle.hooks.firstWhereOrNull(
-      (element) {
-        return p.basename(element.path) == GeneratorHook.postGen.toFileName();
-      },
-    );
-    final pubspec = bundle.hooks.firstWhereOrNull(
-      (element) {
-        return p.basename(element.path) == 'pubspec.yaml';
-      },
-    );
-
-    return GeneratorHooks(
-      preGen: _decodeHookScript(preGen),
-      postGen: _decodeHookScript(postGen),
-      pubspec: _decodeHookPubspec(pubspec),
-    );
-  }
-
-  /// Creates [GeneratorHooks] from a provided [BrickYaml].
-  static Future<GeneratorHooks> fromBrickYaml(BrickYaml brick) async {
-    Future<ScriptFile?> getHookScript(GeneratorHook hook) async {
-      try {
-        final brickRoot = File(brick.path!).parent.path;
-        final hooksDirectory = Directory(p.join(brickRoot, BrickYaml.hooks));
-        final file =
-            hooksDirectory.listSync().whereType<File>().firstWhereOrNull(
-                  (element) => p.basename(element.path) == hook.toFileName(),
-                );
-
-        if (file == null) return null;
-        final content = await file.readAsBytes();
-        return ScriptFile.fromBytes(file.path, content);
-      } catch (_) {
-        return null;
-      }
-    }
-
-    Future<List<int>?> getHooksPubspec() async {
-      try {
-        final brickRoot = File(brick.path!).parent.path;
-        final hooksDirectory = Directory(p.join(brickRoot, BrickYaml.hooks));
-        final file =
-            hooksDirectory.listSync().whereType<File>().firstWhereOrNull(
-                  (element) => p.basename(element.path) == 'pubspec.yaml',
-                );
-
-        if (file == null) return null;
-        return await file.readAsBytes();
-      } catch (_) {
-        return null;
-      }
-    }
-
-    return GeneratorHooks(
-      preGen: await getHookScript(GeneratorHook.preGen),
-      postGen: await getHookScript(GeneratorHook.postGen),
-      pubspec: await getHooksPubspec(),
-    );
-  }
-
-  /// Script run immediately before the `generate` method is invoked.
-  final ScriptFile? preGen;
-
-  /// Script run immediately after the `generate` method is invoked.
-  final ScriptFile? postGen;
-
-  /// Contents of the hooks `pubspec.yaml` if exists.
-  final List<int>? pubspec;
 }
 
 /// {@template generator}
@@ -346,255 +139,6 @@ abstract class Generator implements Comparable<Generator> {
       final response = await http.Client().get(uri);
       return FileContents(target, response.bodyBytes);
     }
-  }
-}
-
-/// File conflict resolution strategies used during
-/// the generation process.
-enum FileConflictResolution {
-  /// Always prompt the user for each file conflict.
-  prompt,
-
-  /// Always overwrite conflicting files.
-  overwrite,
-
-  /// Always skip conflicting files.
-  skip,
-
-  /// Always append conflicting files.
-  append,
-}
-
-/// The overwrite rule when generating code and a conflict occurs.
-enum OverwriteRule {
-  /// Always overwrite the existing file.
-  alwaysOverwrite,
-
-  /// Always skip overwriting the existing file.
-  alwaysSkip,
-
-  /// Always append the existing file.
-  alwaysAppend,
-
-  /// Overwrite one time.
-  overwriteOnce,
-
-  /// Do not overwrite one time.
-  skipOnce,
-
-  /// Append one time
-  appendOnce,
-}
-
-/// {@template directory_generator_target}
-/// A [GeneratorTarget] based on a provided [Directory].
-/// {@endtemplate}
-class DirectoryGeneratorTarget extends GeneratorTarget {
-  /// {@macro directory_generator_target}
-  DirectoryGeneratorTarget(this.dir) {
-    dir.createSync(recursive: true);
-  }
-
-  /// The target [Directory].
-  final Directory dir;
-
-  @override
-  Future<File> createFile(
-    String path,
-    List<int> contents, {
-    Logger? logger,
-    OverwriteRule? overwriteRule,
-  }) async {
-    var _overwriteRule = overwriteRule;
-    final file = File(p.join(dir.path, path));
-    final fileExists = file.existsSync();
-
-    if (!fileExists) {
-      return file
-          .create(recursive: true)
-          .then<File>((_) => file.writeAsBytes(contents))
-          .whenComplete(
-            () => logger?.delayed('  ${file.path} ${lightGreen.wrap('(new)')}'),
-          );
-    }
-
-    final existingContents = file.readAsBytesSync();
-
-    if (const ListEquality<int>().equals(existingContents, contents)) {
-      logger?.delayed('  ${file.path} ${lightCyan.wrap('(identical)')}');
-      return file;
-    }
-
-    final shouldPrompt = logger != null &&
-        (_overwriteRule != OverwriteRule.alwaysOverwrite &&
-            _overwriteRule != OverwriteRule.alwaysSkip &&
-            _overwriteRule != OverwriteRule.alwaysAppend);
-
-    if (shouldPrompt) {
-      logger?.info('${red.wrap(styleBold.wrap('conflict'))} ${file.path}');
-      _overwriteRule = logger
-          ?.prompt(
-            yellow.wrap(
-              styleBold.wrap('Overwrite ${p.basename(file.path)}? (Yyna) '),
-            ),
-          )
-          .toOverwriteRule();
-    }
-
-    switch (_overwriteRule) {
-      case OverwriteRule.alwaysSkip:
-      case OverwriteRule.skipOnce:
-        logger?.delayed('  ${file.path} ${yellow.wrap('(skip)')}');
-        return file;
-      case OverwriteRule.alwaysOverwrite:
-      case OverwriteRule.overwriteOnce:
-      case OverwriteRule.appendOnce:
-      case OverwriteRule.alwaysAppend:
-      case null:
-        final shouldAppend = _overwriteRule == OverwriteRule.appendOnce ||
-            _overwriteRule == OverwriteRule.alwaysAppend;
-        return file
-            .create(recursive: true)
-            .then<File>(
-              (_) => file.writeAsBytes(
-                contents,
-                mode: shouldAppend ? FileMode.append : FileMode.write,
-              ),
-            )
-            .whenComplete(
-              () => shouldAppend
-                  ? logger?.delayed(
-                      '  ${file.path} ${lightBlue.wrap('(modified)')}',
-                    )
-                  : logger?.delayed(
-                      '  ${file.path} ${lightGreen.wrap('(new)')}',
-                    ),
-            );
-    }
-  }
-}
-
-/// A target for a [Generator].
-/// This class knows how to create files given a path and contents.
-// ignore: one_member_abstracts
-abstract class GeneratorTarget {
-  /// Create a file at the given path with the given contents.
-  Future createFile(
-    String path,
-    List<int> contents, {
-    Logger? logger,
-    OverwriteRule? overwriteRule,
-  });
-}
-
-/// {@template script_file}
-/// This class represents a script file in a generator.
-/// The contents should be text and may contain mustache.
-/// {@endtemplate}
-class ScriptFile {
-  /// {@macro script_file}
-  ScriptFile.fromBytes(this.path, this.content);
-
-  /// The template file path.
-  final String path;
-
-  /// The template file content.
-  final List<int> content;
-
-  /// Performs a substitution on the [path] based on the incoming [parameters].
-  FileContents runSubstitution(Map<String, dynamic> parameters) {
-    return FileContents(path, _createContent(parameters));
-  }
-
-  List<int> _createContent(Map<String, dynamic> vars) {
-    try {
-      final decoded = utf8.decode(content);
-      if (!decoded.contains(_delimeterRegExp)) return content;
-      final rendered = decoded.render(vars);
-      return utf8.encode(rendered);
-    } on Exception {
-      return content;
-    }
-  }
-
-  /// Executes the current script with the provided [vars] and [logger].
-  /// An optional [workingDirectory] can also be specified.
-  Future<int> run({
-    Map<String, dynamic> vars = const <String, dynamic>{},
-    Logger? logger,
-    String? workingDirectory,
-    List<int>? pubspec,
-    void Function(dynamic message)? onMessage,
-  }) async {
-    final subscriptions = <StreamSubscription>[];
-    final messagePort = ReceivePort();
-    final errorPort = ReceivePort();
-    final exitPort = ReceivePort();
-
-    var didError = false;
-
-    subscriptions
-      ..add(
-        errorPort.listen((dynamic error) {
-          didError = true;
-          logger?.err('$error');
-        }),
-      )
-      ..add(messagePort.listen(onMessage?.call));
-
-    Uri? packageConfigUri;
-
-    if (pubspec != null) {
-      final directoryHash = sha1.convert(pubspec).toString();
-      final directory = Directory(
-        p.join(Directory.systemTemp.path, '.mason', directoryHash),
-      );
-      final packageConfigFile = File(
-        p.join(directory.path, '.dart_tool', 'package_config.json'),
-      );
-
-      if (!packageConfigFile.existsSync()) {
-        await directory.create(recursive: true);
-        await File(
-          p.join(directory.path, 'pubspec.yaml'),
-        ).writeAsBytes(pubspec);
-
-        final result = await Process.run(
-          'dart',
-          ['pub', 'get'],
-          workingDirectory: directory.path,
-        );
-
-        if (result.exitCode != 0) {
-          logger?.err('${result.stderr}');
-          return ExitCode.software.code;
-        }
-      }
-
-      final packageConfigBytes = await packageConfigFile.readAsBytes();
-      packageConfigUri = Uri.dataFromBytes(packageConfigBytes);
-    }
-
-    final uri = _getHookUri(runSubstitution(vars).content);
-    final isolate = await Isolate.spawnUri(
-      uri,
-      [json.encode(vars)],
-      messagePort.sendPort,
-      paused: true,
-      packageConfig: packageConfigUri,
-    );
-
-    isolate
-      ..addErrorListener(errorPort.sendPort)
-      ..addOnExitListener(exitPort.sendPort)
-      ..resume(isolate.pauseCapability!);
-
-    await exitPort.first;
-    for (final subscription in subscriptions) {
-      unawaited(subscription.cancel());
-    }
-
-    return didError ? ExitCode.software.code : ExitCode.success.code;
   }
 }
 
@@ -762,35 +306,4 @@ List<TemplateFile> _decodeConcatenatedData(List<MasonBundledFile> files) {
   }
 
   return results;
-}
-
-extension on FileConflictResolution {
-  OverwriteRule? toOverwriteRule() {
-    switch (this) {
-      case FileConflictResolution.overwrite:
-        return OverwriteRule.alwaysOverwrite;
-      case FileConflictResolution.skip:
-        return OverwriteRule.alwaysSkip;
-      case FileConflictResolution.append:
-        return OverwriteRule.alwaysAppend;
-      case FileConflictResolution.prompt:
-        return null;
-    }
-  }
-}
-
-extension on String {
-  OverwriteRule toOverwriteRule() {
-    switch (this) {
-      case 'n':
-        return OverwriteRule.skipOnce;
-      case 'a':
-        return OverwriteRule.appendOnce;
-      case 'Y':
-        return OverwriteRule.alwaysOverwrite;
-      case 'y':
-      default:
-        return OverwriteRule.overwriteOnce;
-    }
-  }
 }
