+++ packages/mason_cli/test/mason_generator_test.dart
@@ -70,11 +70,7 @@ void main() {
         const vars = {'name': 'Dash'};
         final target = Directory.systemTemp.createTempSync();
 
-        await hooks.preGen?.run(
-          vars: vars,
-          logger: logger,
-          workingDirectory: target.path,
-        );
+        await hooks.preGen?.run(vars: vars, workingDirectory: target.path);
         expect(
           File(path.join(target.path, '.pre_gen.txt')).existsSync(),
           isTrue,
@@ -91,11 +87,7 @@ void main() {
           isTrue,
         );
 
-        await hooks.postGen?.run(
-          vars: vars,
-          logger: logger,
-          workingDirectory: target.path,
-        );
+        await hooks.postGen?.run(vars: vars, workingDirectory: target.path);
         expect(
           File(path.join(target.path, '.post_gen.txt')).existsSync(),
           isTrue,
